methodes*3~4 pour la classe recette

+++ src/main/kotlin/org/ldv/savonapi/model/entity/Resultat.kt
@@ -8,7 +8,7 @@ class Resultat(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var resultatId: Long,
-    var score: Long,
+    var score: Float,
 
     @ManyToOne
     @JoinColumn(name ="recetteId")
